==> analyses/molecular-subtyping-LGAT/02-subset-fusion-files-LGAT.Rmd <==
---
title: "Annotate Fusion defining subtype status for LGAT biospecimens"
output: html_notebook
author: K S Gaonkar 
date: 2020
---

As per [issue](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/790) we will be subtyping LGAT based on fusion  in the following genes:

 - LGG, KIAA1549-BRAF
contains KIAA1549-BRAF fusion

 - LGG, other MAPK
contains non-canonical BRAF fusion other than KIAA1549-BRAF
contains RAF1 fusion

- LGG, RTK
harbors a fusion in ALK, ROS1, NTRK1, NTRK2, or NTRK3 or
harbors a PDGFRA fusion

- LGG, FGFR
harbors FGFR1-TACC1 fusion 
harbors FGFR1 or FGFR2 fusions

- LGG, MYB/MYBL1
harbors either a MYB-QKI fusion or other MYB or MYBL1 fusion


### Setup
```{r}
library(tidyverse)

# Look for git root folder
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))

# get subset folder
subset_dir <- file.path(root_dir, 
                        "analyses",
                        "molecular-subtyping-LGAT",
                        "lgat-subset")

# create if doesn't exist
if (!dir.exists(subset_dir)) {
  dir.create(subset_dir)
}
```

### Input

**LGAT fusion subset file**

`fusion_summary_lgat_foi.tsv` was generated by `fusion-summary` module for LGAT RNA biospecimens.

 - all fusions in the above gene of interest per subtype was summarized "1" for present and "0" if absent.
 
 - for kinase fusions, the following conditions needed to be satisfied for LGAT as per [PR](https://github.com/AlexsLemonade/OpenPBTA-analysis/pull/830):
 
 1. Added all 3' kinase fusions which are in-frame and retain the kinase domain
 2. For 5' kinase fusions, added those which are in-frame and retain the kinase domain
 3. For 5' kinase fusions that are not in-frame (from point 2 above), check whether it has a reciprocal fusion and whether that reciprocal is in-frame and the kinase domain is retained. Then, add those 5' kinase fusions.


```{r}

lgat_specimens_df<-read_tsv(file.path(subset_dir, "lgat_metadata.tsv"))

# Filter to RNA-Seq samples
lgat_rna_df <- lgat_specimens_df %>%
  dplyr::filter(experimental_strategy == "RNA-Seq") %>%
  dplyr::select(Kids_First_Biospecimen_ID,sample_id)

# get putative oncogene fusion list
putativeFusion <- readr::read_tsv(file.path(root_dir,
                                            "data",
                                            "fusion_summary_lgat_foi.tsv")) %>%
  # add sample_id for LGAT RNA-Seq biospecimen
  dplyr::left_join(lgat_rna_df, by=c("Kids_First_Biospecimen_ID")) %>%
  # filter to keep only LGAT RNA 
  dplyr::filter(!is.na(sample_id))
```


Gather gene(s) that define LGAT subtypes
Additional information for canonical fusion for each subtype is also provided in the list. 

```{r}
# fusion list of interest
fusionOI <- jsonlite::fromJSON(file.path(root_dir,
                                         "analyses",
                                         "molecular-subtyping-LGAT",
                                         "input",
                                         "fusionOI_list.json"))

# collapse gene list with "|" for easier match()
MAPK_fused_gene <- paste(fusionOI$MAPK$gene,collapse = "|")
RTK_fused_gene <- paste(fusionOI$RTK$gene,collapse = "|")
FGFR_fused_gene <- paste(fusionOI$FGFR$gene,collapse = "|")
MYB_fused_gene <- paste(fusionOI$MYB$gene,collapse = "|")
canonical_fusion <- paste(fusionOI$canonical$fusion,collapse = "|")
```


### Format and annotate biospecimen per subtype based on Fusion
We will gather summarized fusion calls that satisfy the conditions per subtype and save the subset

```{r}
subsetFusion <- putativeFusion %>%
  mutate(
    # get KIAA1549--BRAF status
    KIAA_BRAF_fus = case_when(
      # canonical BRAF fusion 
      rowSums(dplyr::select(putativeFusion,dplyr::matches(canonical_fusion))) > 0 ~ "Yes",
      TRUE ~ "No"
    ),
    # other MAPK fusion status
    MAPK_fus = case_when(
      # Fusion with BRAF and RAF1 
      rowSums(dplyr::select(putativeFusion,dplyr::matches(MAPK_fused_gene))) > 0 &
        # remove biospecimens with canonical fusion
        # they are a separate subtype as shown above
        !rowSums(dplyr::select(putativeFusion,dplyr::matches(canonical_fusion))) > 0 ~ "Yes",
      TRUE ~ "No"
    ),
    # RTK fusion status
    RTK_fus = case_when(
      # fusion in any RTK gene ALK|ROS1|NTRK2|NTRK1|NTRK3|PDGFRA
      rowSums(dplyr::select(putativeFusion,dplyr::matches(RTK_fused_gene))) > 0 ~ "Yes",
      TRUE ~ "No"
    ),
    # get FGFR fusion status
    FGFR_fus = case_when(
      # fusion in any FGFR gene FGFR1|FGFR2
      rowSums(dplyr::select(putativeFusion,dplyr::matches(FGFR_fused_gene))) > 0 ~ "Yes",
      TRUE ~ "No"
    ),
    # get MYB mutation status
    MYB_fus = case_when(
      # fusion in any MYB gene MYB|MYBL1
      rowSums(dplyr::select(putativeFusion,dplyr::matches(MYB_fused_gene))) > 0 ~ "Yes",
      TRUE ~ "No"
    )
  ) %>%
  dplyr::select(Kids_First_Biospecimen_ID,
                sample_id,
                KIAA_BRAF_fus,
                MAPK_fus,
                RTK_fus,
                FGFR_fus,
                MYB_fus)

# save to subset folder
write_tsv(subsetFusion,file.path(subset_dir, "LGAT_fusion_subset.tsv"))
```


```{r}
sessionInfo()
```
